log download activity when using cloudstorage
